backend/requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.5
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.7
pyjwt==2.8.0
bcrypt==4.0.1
